going a different way

+++ server/lib/pyasn1/type/univ.py
@@ -425,7 +425,7 @@ class Null(OctetString):
     subtypeSpec = OctetString.subtypeSpec+constraint.SingleValueConstraint(''.encode())
 
 if sys.version_info[0] <= 2:
-    intTypes = (int, int)
+    intTypes = (int, long)
 else:
     intTypes = int
